Split trailing null values logic into 2 files
