+++ packages/app/src/pages/landelijk/intensive-care-opnames.tsx
@@ -34,9 +34,9 @@ import {
 } from '~/static-props/get-data';
 import { colors } from '~/style/theme';
 import { IntakeHospitalPageQuery } from '~/types/cms';
+import { countTrailingNullValues } from '~/utils/count-trailing-null-values';
 import { getBoundaryDateStartUnix } from '~/utils/get-trailing-date-range';
 import { replaceVariablesInText } from '~/utils/replace-variables-in-text';
-import { countTrailingNullValues } from '~/utils/trailing-null-values';
 
 export const getStaticProps = createGetStaticProps(
   getLastGeneratedDate,
